// src/app/@restapi/model/user.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface User {
    authorized?: Array<string>;
    created?: string;
    email?: string;
    firstName?: string;
    lastName?: string;
    modified?: string;
    password?: string;
    phone?: string;
}